+++ scripts/library_preparation/protonation/protgen_GypsumDL.py
@@ -14,7 +14,7 @@ from rdkit.Chem import PandasTools
 dockm8_path = next((p / 'DockM8' for p in Path(__file__).resolve().parents if (p / 'DockM8').is_dir()), None)
 sys.path.append(str(dockm8_path))
 
-from scripts.utilities import parallel_executor, printlog, split_sdf_str
+from scripts.utilities.utilities import parallel_executor, printlog, split_sdf_str
 
 warnings.filterwarnings("ignore", category=UserWarning)
 warnings.filterwarnings("ignore", category=DeprecationWarning)
